requirements.txt
pynput>=1.7.6
Pillow>=10.0.0
pywin32>=306
psutil>=5.9.0
flask>=3.0.0
flask-cors>=4.0.0
cryptography>=41.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
plyer>=2.1.0
opencv-python>=4.8.0
face-recognition>=1.3.0
requests>=2.31.0
